+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * Prevent a problem in which if an exposure fails with an uncaught error the listener callback was not removed, which would cause duplicate messages being output in successive exposures.
 * Try and except when saving exposures to disk.
+* Fix case in actor when camera fails to return status.
 
 
 ## 0.5.4 - January 7, 2022